improve memory usage and allocation
Do not realloc when it is not needed (even when it is the same size).
Decrease the greedy allocation increment size for nested nodes also.

Tested for example using valgrind and "add Beej's Guide to Network Programming" HTML page:
https://git.codemadness.org/webdump_tests/commit/837749abc02f28e1584e5f2cf2b274ae1c69d8e6.html

The buffering for link references (-l option) used way too much memory.